third_party/WebKit/LayoutTests/http/tests/worklet/usecounter-on-addmodule.html

<!DOCTYPE html>
<title>Worklet: UseCounter on Worklet.addModule()</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>

// This test should not be upstreamed to WPT because this tests the UseCounter
// that is not defined in Web standards.

// WebFeature::kWorkletAddModule defined in web_feature.mojom.
const kFeature = 2364;

test(() => {
  assert_false(internals.isUseCounted(window.document, kFeature));
  CSS.paintWorklet.addModule('resources/empty-worklet-script.js');
  assert_true(internals.isUseCounted(window.document, kFeature));
}, 'The usage count of Worklet.addModule() should be recorded in UseCounter.');

</script>
